+++ packages/devtools/cli/src/util/publish/common.ts
@@ -6,7 +6,6 @@ import { ux } from '@oclif/core';
 import { CID } from 'ipfs-http-client';
 
 import type { ConfigProto } from '@dxos/config';
-import { TunnelResponse } from '@dxos/protocols/proto/dxos/service/publisher';
 
 export type PackageModule = NonNullable<NonNullable<ConfigProto['package']>['modules']>[0];
 export type PackageRepo = NonNullable<NonNullable<ConfigProto['package']>['repos']>[0];
@@ -19,14 +18,6 @@ export const mapModules = (modules: PackageModule[]) => {
   }));
 };
 
-export const mapTunnels = (tunnels: TunnelResponse[]) => {
-  return tunnels.map((tunnel) => ({
-    key: tunnel.name,
-    enabled: tunnel.enabled,
-    url: tunnel.url,
-  }));
-};
-
 export const printModules = (modules: PackageModule[], flags = {}) => {
   ux.table(
     mapModules(modules),
@@ -36,26 +27,6 @@ export const printModules = (modules: PackageModule[], flags = {}) => {
       },
       bundle: {
         header: 'Bundle',
-      },
-    },
-    {
-      ...flags,
-    },
-  );
-};
-
-export const printTunnels = (tunnels: TunnelResponse[], flags = {}) => {
-  ux.table(
-    mapTunnels(tunnels),
-    {
-      key: {
-        header: 'App',
-      },
-      enabled: {
-        header: 'Enabled',
-      },
-      url: {
-        header: 'URL',
       },
     },
     {
