+++ packages/devtools/cli/src/util/tunnel/common.ts
@@ -1,0 +1,35 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import { ux } from '@oclif/core';
+
+import { TunnelResponse } from '@dxos/protocols/proto/dxos/service/tunnel';
+
+export const mapTunnels = (tunnels: TunnelResponse[]) => {
+  return tunnels.map((tunnel) => ({
+    key: tunnel.name,
+    enabled: tunnel.enabled,
+    url: tunnel.url,
+  }));
+};
+
+export const printTunnels = (tunnels: TunnelResponse[], flags = {}) => {
+  ux.table(
+    mapTunnels(tunnels),
+    {
+      key: {
+        header: 'App',
+      },
+      enabled: {
+        header: 'Enabled',
+      },
+      url: {
+        header: 'URL',
+      },
+    },
+    {
+      ...flags,
+    },
+  );
+};
