+++ packages/devtools/cli/src/util/tunnel/index.ts
@@ -2,4 +2,5 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './common';
 export * from './tunnel-rpc-peer';
